Combined updateWinery and deleteWinery actions into one action on the UpdateWinery.jsx component. All working. This will allow to add create or delete review combined action in WinerySingle.jsx.

+++ client/src/utils/actions/winery.js
@@ -12,22 +12,51 @@ export async function createWinery(request) {
   })
 }
 
-export async function updateWinery(request, id){
-  const data = await formToObj(request)
-  return await axios.put(`/api/wineries/${id}`, data, {
-    validateStatus: () => true,
-    headers: {
-      Authorization: `Bearer ${getToken()}`
-    }
-  })
+// export async function updateWinery(request, id) {
+//   const data = await formToObj(request)
+//   return await axios.put(`/api/wineries/${id}`, data, {
+//     validateStatus: () => true,
+//     headers: {
+//       Authorization: `Bearer ${getToken()}`
+//     }
+//   })
+// }
+
+// export async function deleteWinery(id) {
+//   await axios.delete(`/api/wineries/${id}`, {
+//     validateStatus: () => true,
+//     headers: {
+//       Authorization: `Bearer ${getToken()}`
+//     }
+//   })
+//   return redirect('/wineryIndex')
+// }
+
+export async function updateOrDeleteWinery(request, id) {
+  // console.log(request)
+  // console.log(id)
+  let data = await formToObj(request)
+  // setTimeout(console.log(data), 5000)
+
+  if (data.intent === 'update') {
+    return await axios.put(`/api/wineries/${id}`, data, {
+      validateStatus: () => true,
+      headers: {
+        Authorization: `Bearer ${getToken()}`
+      }
+    })
+  }
+  if (data.intent === 'delete') {
+    await axios.delete(`/api/wineries/${id}`, {
+      validateStatus: () => true,
+      headers: {
+        Authorization: `Bearer ${getToken()}`
+      }
+    })
+    return redirect('/wineryIndex')
+  }
 }
 
-export async function deleteWinery(id) {
-  await axios.delete(`/api/wineries/${id}`, {
-    validateStatus: () => true,
-    headers: {
-      Authorization: `Bearer ${getToken()}`
-    }
-  })
-  return redirect('/wineryIndex')
+export async function createOrDeleteReview(request, id) {
+  
 }
